Fixes #16013: Add benchmarks for big nodes lists
